Fix the bug that vti reported when the new 304 caching code played badly with 404 handling.
The solution is to serve 404s specially, so that they always succeed in setting
404, and never return anything else, like 304 or 403. We do this in part by
using internal 404 pages if the physical 404 page requested is not available.

Also a default location for 404 files is provided so it doesn't have to set
every time.

This patch adds a new test script, the first one for testing
MojoX::Dispatcher::Static directly.