Document 'cookies()' in Mojo::Message::Request.
It would be very convenient if this was also supported:

 $first_cookie_obj  = $req->cookie('my_name');
 @all_cookie_objs   = $req->cookie('my_name');

Commonly we just want to deal a cookie by a specific name, and don't care if
it's the 1st or the 5th cookie returned by 'cookies().  CGI.pm, CGI::Simple and
Catalyst and CGI::Application all have an option to retrieve a cookie by name.